fix(dawcore): extract shared viewport util, permissive controller defaults
- Extract getVisibleChunkIndices to utils/viewport.ts (shared by
  ViewportController and daw-waveform, eliminates duplication)
- Initialize ViewportController with visibleStart=-Infinity, visibleEnd=Infinity
  so all chunks render before attachScrollContainer is called
- daw-waveform._getVisibleChunkIndices delegates to shared function
